CMP coverage

+++ coverage.cpp
@@ -182,13 +182,13 @@ void WriteCoverageBinary(Coverage& coverage, FILE *fp) {
     fwrite(iter->module_name, sizeof(iter->module_name), 1, fp);
     uint64_t num_offsets = iter->offsets.size();
     fwrite(&num_offsets, sizeof(num_offsets), 1, fp);
-    uint64_t *offsets = (uint64_t *)malloc(num_offsets * sizeof(uint64_t));
+    uint64_t *offsets = (uint64_t *)malloc((size_t)num_offsets * sizeof(uint64_t));
     size_t i = 0;
     for (auto iter2 = iter->offsets.begin(); iter2 != iter->offsets.end(); iter2++) {
       offsets[i] = *iter2;
       i++;
     }
-    fwrite(offsets, sizeof(uint64_t), num_offsets, fp);
+    fwrite(offsets, sizeof(uint64_t), (size_t)num_offsets, fp);
     free(offsets);
   }
 }
@@ -211,8 +211,8 @@ void ReadCoverageBinary(Coverage& coverage, FILE *fp) {
     fread(module_coverage.module_name, sizeof(module_coverage.module_name), 1, fp);
     uint64_t num_offsets;
     fread(&num_offsets, sizeof(num_offsets), 1, fp);
-    uint64_t *offsets = (uint64_t *)malloc(num_offsets * sizeof(uint64_t));
-    fread(offsets, sizeof(uint64_t), num_offsets, fp);
+    uint64_t *offsets = (uint64_t *)malloc((size_t)num_offsets * sizeof(uint64_t));
+    fread(offsets, sizeof(uint64_t), (size_t)num_offsets, fp);
     for (size_t i = 0; i < num_offsets; i++) {
       module_coverage.offsets.insert(offsets[i]);
     }
